requirements.txt
fastapi[all]
uvicorn
pydantic
python-dotenv
PyPDF2
google-generativeai
langchain-google-genai
langchain